+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Added
 
 - 単価オファーに事前予約要件を追加
+- 単価オファー検索条件拡張
 
 ### Changed